requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
langchain
openai
faiss-cpu
sentence-transformers
streamlit
tqdm
